[story_shell] Add a test for story shell.
This test adds one Module, then another Module, using Intents to
a running story. Then stops the story and starts it again and
verifies that the modules get added with their module manifests
passed to the story shell.

This uncovers a crapload of bugs in the daisy based AddModule() code:

* AddModule() allows null surface relations, but internally null surface
  relation values are used to represent embedded modules.

* A null parent module in AddModule() still assigns the first module
  as display parent.

* If AddModule() is called before the Module runs, the module is started
  anyway, but the daisy is not saved. AddModule() was implemented with
  total disregard to the possibility that it's called before the
  story runs.

To be fixed in followup changes.

MS-1473 #comment Add test for story shell.

Change-Id: I98018cdfa195cddff923db40478fb38b3c6bdf09