[blobfs] Implement optional blobfs compression
Consider compression during writeback, and only
write to disk in compressed format if a minimal
number of blocks would be saved.

Overview:
- Compression utilities are implemented in lz4.h / lz4.cpp.
- Compression is aborted partway through writeback
if not enough storage would be saved, avoiding
the overhead of needless compression.
- Storage space for compressed buffer is released
immediately when the blob write has completed.
- Decompression happens up front in "InitVmos".
- Since writeback may be delayed, paginate it for large
buffers to avoid overflowing the writeback buffer. Add
tests for this condition in both the compressed and
uncompressed cases.

ZX-2035 #done

Change-Id: I231190bde0939094e932aa1eb28281e3811aca56