[dev][block][bootpart] Add support for bootdata partition maps
A new driver bootpart (based on the GPT driver) supports partition maps
attached via device metadata.
This allows partitioning raw block devices using a partition map
provided by the bootloader.
This is needed for devices that do not contain a standard partition map
like GPT or MBR embedded in the storage medium.

A new block_info_t flag BLOCK_FLAG_BOOTPART indicates the presense of a
bootdata partition map. This is used by the block watcher to bind the
bootpart driver to block devices that have a bootdata partition map.

Change-Id: Ie00ee8d4ec190a2c95fd6e668f067bc4bcc2fbf3