[ulib][unittest] Add watchdog timer
If the timer goes off before the test completes then the process is terminated.

A new option --w <seconds> is added to runtests to set the timeout.
A value of zero disables the timeout.

This is passed to tests via the RUNTESTS_WATCHDOG_TIMEOUT env var.

New -w option manually tested with:
$ runtests -w
$ runtests -w ""
$ runtests -w -1
$ runtests -w 123foo
$ runtests -w 2147483648
$ runtests -w 0x80000000
All flag an error as expected, and these work:
$ runtests -w 2147483647
$ runtests -w 0x7fffffff
And this works and triggers timeouts:
$ runtests -w 2

ZX-2018 #done

Change-Id: Ia9bc9e29ce93da3973ac58f63393096b5c836d37