[netstack] Update bitfields in netstack FIDL interface to use constants
Enums in FIDL are awkward for bitfield entries since they generate enum
class types in C++ that do not implicitly convert to integer values,
requiring awkward casts to manipulate. This modifies the netstack
interface feature set to be expressed as a set of constants instead of
enum values.

FIDL-164 covers having more generic guidance and tooling around
bitfields in FIDL.

Change-Id: I34b4fb4916f89485e01e29fc356ce3031c18e754