[audio][mixer] Refactor MuteThreshold()
Because our gain stages are pre-SUM (done per-stream, not per-
output), and because all input formats are normalized to the
same maximum value, the value for Gain::MuteThreshold is the
same for all inputs, regardless of input format's bitwidth.

With this in mind, this CL reworks Gain::MuteThreshold to
remove this parameter. In the process, it removes from callers
any notion of our processing pipeline width.

This CL is a precursor to incoming work that will increase the
width of the audio processing pipeline.

Change-Id: I472d9cb495fbeb206bc51261b1922a88c413c840