[media_player] video renderer tweaks
This CL improves the video player a bit by changing the way it
handles incoming packets. The changes improve player performance
slightly when CPU-constrained.

The most important change is that AdvanceReferenceTime is called
on packet arrival. This allows the renderer to dismiss 'stale'
packets earlier than it otherwise would, making it more likely
that 'fresh' packets are available when we're ready for them.

In addition, logic involving the completion of prime requests is
handled more proactively. Previously, the code would wait longer
than required to declare that priming is complete.

Change-Id: If520cc7b16169f200463f7d643cd54ec5826d8ad