[fidl] Have C++ clients log errors when writing to channel.
Ideally we'd return the zx_status_t to the client and let them deal with
it, but the code is currently structured such that the client proxy
class is a sub-class of the service interface (class {{ .Name }} in the
bindings generation code). And it doesn't make sense for the real server
implementation (which implements that same interface) to return non-void.

Change-Id: I5325825412ff018979210cf4e38083ba85600d09