[kernel] Don't leak threads that haven't started
Previously, if you created a thread but never started the thread, the
ThreadDispatcher would AddRef() itself but never Release() itself,
causing a leak. The Release() is called by ThreadDispatcher::Kill(),
which isn't called until the thread is added to the threadlist of its
process (which doesn't happen until the thread actually starts).

ZX-2049 #done

Change-Id: I6b836b5456b4cb8a981c7b870666ea9a94425733