[story_shell] Don't arbitrarily reparent top level modules.
The hack to pick an arbitrary display anchor for an external top
level module doesn't work, because it does not preserve order in
which external modules are presented to the story shell when a
story resumes. We have fixed the behavior that relied on this
hack elsewhere, and now remove the hack altogether.

MS-1473 #comment  Don't arbitrarily reparent top level modules.

Change-Id: Ia9cc87e4ceeb638d36a304feadfee7fefbf3db50