[debugger] Start linking with LLVM
Add a LLVM disassembler wrapper. This is not hooked up to the command
line tool yet but is verified with tests. Only x64 is hooked up.

Stop distributing zxdb as a target tool, remove packaging. I decided it
will be too annoying to maintain a target one without LLVM code.

Add LLVM state to the session for talking about the target system. Start
linking to LLVM.

Fix the message loop test on Linux (it was using the wrong ends of the
pipes but apparently Zircon doesn't care). This makes all zxdb tests
pass on the host.

Add groups "host_tools" and "host_tests" to the Garnet build so the zxdb
files are referenced, and so there is an easier way to build them.

The OutputBuffer class is moved to the client directory so that the
disassembler can use it to emit syntax-colored output.

Change-Id: I3bc805fa73b913599cd2ddfae47ff13657d38ee7