[runbenchmarks] Write summary.json before exiting
summary.json is required to run perf tests on hardware.  The
infrastructure uses this as a manifest to figure out which files
should be copied off of the hardware device.  It additionally
reports which tests passed or failed.

Example summary.json generated at this change for garnet benchmarks:

  {"tests":[
  {"name":"benchmark_example","output_file":"benchmark_example","result":"PASS"},
  {"name":"zircon_benchmarks","output_file":"zircon_benchmarks","result":"PASS"}
  ]}

I'm using this summary.json from a past garnet build as a guideline:
https://logs.chromium.org/v/?s=fuchsia%2Fbuildbucket%2Fcr-buildbucket.appspot.com%2F8948348994252747072%2F%2B%2Fsteps%2Ftest_results%2F0%2Flogs%2Fsummary.json%2F0

IN-327

Change-Id: I738dd355a2b3b9d8c351aa21eaceb2cba0d622d4