[test] Make run_integration_tests return error on missing config file.
run_integration_tests states that a configuration file is required but
exits with code 0 even if it's not present. This means a misconfigured
test setup could appear to succeed even though no tests are run. This
cl changes run_integration_tests to exit with 1 if the configuration
file is missing.

IN-332 #comment The final cl for this change.

Change-Id: I19a458f64ac1b6ced7e33b69f21ea49923d5cb6f