[sdk] Add standard library implementation SOs to dist
This adds the implementation libraries for libc++.so.2, libc++abi.so.1
and libunwind.so.1 to the arch/{x64,arm64}/dist/ directories of the
SDK. These need to be packaged by users of libfxl.so, etc. The usual
SDK packaging mechanism doesn't identify these libraries as our build
locates them in specialized logic:

https://fuchsia.googlesource.com/build/+/master/gn/finalize_manifests.py#289

that the generic mechanism is currently unaware of.

DX-135 #done

Change-Id: I8ba2313bc23fc6bbce85801d95adab4b39218ca0