[zxcrypt] Don't cache parent block information
This CL removes the copy of the block_info_t from the zxcrypt
device.  Caching such information caused it to be invalid after FVM
resizes.  This also caused the 'synthetic' block size to be removed,
which hasn't been needed since the Jan. refactor, and which broke
atomicity with lower layers.

ZX-1948 #comment Fixed

Change-Id: I8d2e9ab8eb965f7ead1796eb38d9dc0e6ee939f8