[audio][mixer] Fix throttle output behavior.
Fix two issues which were effecting the throttle output's ability to
release packets in a timely fashion.

First, audio renderers currently hold a reference to the throttle
output so that they can prime new outputs which show up when a routing
event occurs.  In addition the throttle output also is
always supposed to exist in the set of active outputs for any
renderer.  Renderers were explicitly queueing packets to the throttle
output in addition to queueing packets to each of their active outputs
every time a packet arrived.  This was causing the packets to exist in
the throttle output queue two times when they should have been their
only once.

While incorrect, this should normally only be inefficient, not a
problem.  The second issue showed up with an optimization which was
made in 917c2a75f925af737df12ad7074e57ad02afd1ed.

During ForeachLink in the standard output, after processing the head
of a source's packet, queue, the optimization checking to see if it
had produced all of the frames it needed to for this mix cycle, and
fast aborting the processing of the source if it had.  The problem
with this is that Trim cycles *never* have any frames to produce, so
this could would *always* fast abort.  Thankfully, the code which
followed would pay attention to the signal which came from ProcessTrim
and trim the buffer from the front of the pending queue.
Unfortunately, it would only ever consider a single packet per cycle,
and never any more.  This meant that if the packets were coming in at
a rate faster than once every 10mSec (the current rate that the
throttel output checks for trim), that the throttle output would just
become hopelessly backlogged and never catch up.

The fix is to only pay attention to this fast abort path during a mix
cycle (when we expect to produce frames) and never during a trim
cycle.

Change-Id: Ibf61fd61ce75518413ab7990ba88a565683a7811