[examples][tones] Make tones obey min lead time
Update the tones example to take into account the minimum lead time
requirement reported by its renderer.

Note, it is still not paying attention to updates to the minimum lead
time, so if the min lead time requirement grows due to a routing
change while the example is running, the program will need to be
restarted in order to take this into account.

Eventually, if we move this to a model driven purely by timing
(instead of packet completion) in order to reduce latency, it should
be easier to adapt to lead time requirement changes on the fly.

Change-Id: Icb1cdbc37f858b69f58976b977058a10f1c8cf58