[examples][tones] Fix frequency mappings.
The help for the interactive mode of the tones example was showing
note ordinal zero being bound to a "C", on a piano keyboard, but was
mapping the note to an "A"; specifically A2 (110 Hz).

Change this so that ordinal zero now maps to C4, or middle C on a
standard piano keyboard.

Change-Id: I69a88c925a3bf13ebb51871c5aef8336469615c9