[kernel][x86] Correct stride for pair on relocation fixup
Ensure a pair use the correct stride instead of defaulting to 8. This
bug was not found with mawk because the size check (< 0x1000) was not
interpreted correctly. The relocation is mute on default but the
issue is visible on the kaslr static relocation change I am working on.

SEC-31 #comment Correct bug on x86 pair fixup that was hidden when the script was using mawk

Change-Id: Ied08e2278f4e1b12aa595917b17f0753b7645469