[kernel][vm] provide a generic routine for allocating kernel stacks
Move a specialized routine out of the ThreadDispatcher into a common
place for allocating a kernel stack out of a vmar + mapping with a guard
page in front of and behind.

Chang the arm generic code to use this for cpu bootup instead of
allocating stacks out of the physmap.

Change-Id: Id38c8e270ae9fc365b200bad89265f320fa70ab2