[media_player] clean up framework node model naming
The in-proc streaming framework supports a number of models for
nodes in the graph. This CL eliminates one model and renames some
of the others.

Sources and sinks (nodes that have no inputs and no outputs,
respectively) are almost always 'active' (asynchronous), so there's
no point in having Active in the name. If we have synchronous sources
and sinks in the future, they can easily be accommodated with the
asynchronous models. The difference is not so great, and there's not
much of a performance win to having separate synchronous models.
Accordingly, the old MultistreamSource, a synchronous model, has been
deleted and ActiveMultistreamSource has been renamed over it.

In the near future, there will be a new model, AsyncTransform. The
operation of an asynchronous transofrm is sufficiently different from
synchronous operation that a separate model is warranted.

Change-Id: I081d05f180938a5fe0869f0c0aef4e5973241a03