[machina][qcow] Add refcount read/write component.
The QCOW refcount table tracks how physical (host) clusters are
allocated. Any cluster with a refcount greater than 0 is currently
in use, either holding data clusters or containing other metadata, such
as the L1/L2 tables and the refcount tables.

This will be used in a following change to allocate unused clusters as a
part of supporting writes.

Change-Id: I080d1106f422bef3ebdd8adff312140d6cc7cf31