[kernel][futex] Use resched_disable instead of passing around "any_woken"
Change the futex implementation to use resched_disable instead of
manually tracking whether we need to call thread_reschedule().

Add an RAII helper class, AutoReschedDisable.

This shouldn't change the performance of futexes.  This is a cleanup
in the sense that it removes a use of "reschedule=false" and
simplifies the futex code.  Adding AutoReschedDisable will be useful
for later changes which will improve performance.

ZX-1690

Change-Id: Ieab1a39ddf6a55ad049fe9b95ddb9884a3b67265