[wlan] wlantap driver stub, attempt #2
Implement a wlantapctl device that listens to ioctls and creates fake
wlan devices.

Currently, the driver simply prints lines in the log instead of sending
messages back to the client. That part is yet to be figured out.

This time, as a workaround for ZX-2064, we bind to /dev/test instead
of /dev/misc. The only other difference from the original change is the
MAC role field coversion in wlantap-mac.cpp, to make it work with
https://fuchsia-review.googlesource.com/c/garnet/+/145007.

NET-660

This is a revert of a revert
https://fuchsia-review.googlesource.com/c/garnet/+/147338

Change-Id: I8732021f71ec4b1ff4cb5a14ee0a180fdc503fec