[debugger] Async connect, add arch.
For symbol and disassembly support the client needs to know the
architecture of the target system. This patch adds an enum on the
"hello" message reply with that information.

Since this information is now required, the "hello" message is necessary
before a connection can be used, which requires connection be
asynchronous. In addition to the asynchronous "hello" message, the DNS
and socket calls are now done on a background thread, and there is a way
to cancel pending connections (previously "connect" would just hang if
the socket wasn't responding, forcing you to kill the process).

Removes the restriction on the MessageLoop that watches not be removed
from the callback since this patch hit the issue.

An additional signature check is added to the hello reply to help verify
the remote end is the correct service. Previously it would just be
confused as it tried to deserialize a message with a random sized header
and hang.

Change-Id: Ief87d724b2b09080135739c0462085e2e7bc15c1