[zircon][amlogic][emmc]Make interrupts and Descriptor chaining work
This changeset enables use of the interrupts and descriptor chaining for
emmc. The descritptor chaining follows the sdhci controller descriptor chaining
mainly, but doing things differently for amlogic controller as required.
For example, amlogic controller does not have a fifo data buffer like
sdhci controller in its registers. So we allocate a dma data buffer
even when the upper layers send a sdmmc request with use_dma flag set to
false.

Change-Id: I9f43a7ecb1d79b69ec93857d4581e330ee97c581