[pci] Handle IRQ modes from any existing state
Allow a driver to request an IRQ mode while in any existing IRQ
mode provided there are no handlers presently configured for the
IRQ in question. This handles the case where a driver dies or is
killed after configuring interrupts and finds the devicein a bad
state when a new driver is spawned.

Change-Id: I47ccc5d19a4d8addcbce2a81dbd6efc4a8248001