[debugger] Hook up a disassemble command
This command requests memory using the existing infrastructure, and
uses the newer disassembler to decode it.

Adds support to the disassembler to handling MemoryDumps, which are more
tricky because they can have invalid regions.

Move the zxdb_test_app to the debug_agent directory since the zxdb
directory is no longer referenced from the target build.

Change-Id: I97e480d4da2425be794561ecd8b4e4f1f66b6c28