[fx] Pass the same -j option to build-zircon.sh and Ninja
The Zircon build now supports GOMA, and out/.../fx.config will
cause GOMACC to be set in the environment for fx commands so that
Zircon will use GOMA by default if GN is using it.  So the Zircon
build should get the same level of concurrency as the Ninja build.

Change-Id: Ib1095e0b08306120be20be1ce4286e8debd30153