[kernel] Guard against potential recursive deadlock
user_callback calls ThreadDispatcher, which (eventually) calls
thread_reschedule, which will acquire the thread scheduler lock.

Use a thread annotation to help ensure scheduler lock is not held when
invoking thread's user_callback.  See ZX-2055 for background.  Of
course, this is not foolproof as there can be false negatives with
TA_EXCL.

ZX-2055

Change-Id: I7c523cf4f4513ece72211b43530b416a1d32c8b4