[fx] Remove board argument to fx boot
The binaries used for booting are identical for both x64 and arm64
so the board argument is no longer necessary.
However the build system has not been updated to take advantage of that.
So for now, we use board "pc" for x64 and "qemu" for all arm64 targets.
After the build system has been updated we can remove the remaining usage of ${board}

Change-Id: I5f99888d530178a782c30d5517dad4378e8d1eaa