[kernel][port] Use resched_disable instead of tracking a wake_count
Change PortDispatcher and Semaphore to use resched_disable instead of
manually tracking whether we need to call thread_reschedule().

This shouldn't change the performance of ports.  This is a cleanup in
the sense that it removes a use of "reschedule=false" and simplifies
the code.

Remove the !arch_ints_disabled() assertion from
thread_resched_disable(), because PortDispatcher::Queue() is called
with interrupts disabled via the following code path, which sends a
debug message when a thread exits as a result of being killed:

  x86_iframe_process_pending_signals()/arm64_thread_process_pending_signals()
  thread_process_pending_signals()
  thread_do_suspend()
  check_kill_signal()
  thread_exit()
  ThreadDispatcher::Exiting()
  ExceptionPort::OnThreadExitForDebugger()
  ThreadDispatcher::ExceptionHandlerExchange()
  ExceptionPort::SendPacketWorker()
  PortDispatcher::Queue()

The assertion was unnecessarily strict.  Note that there are very few
places in the kernel that assert that interrupts are *not* disabled.

ZX-1690

Change-Id: I8142295202bfce332abb69e2b36a79bda0df20ab