[triggers] Add message queue trigger for queue deletion
This allows an agent to observe the deletion of any message queue
that it knows the token for. This is accomplished by introducing
a new trigger condition for tasks.

The reason for using the message queue token as opposed to the name
(as is done for new message triggers), is to allow agents to observe
the deletion of queues that it does not know the name for (i.e.
didn't create themselves).

MI4-946 # comment [triggers] Add message queue trigger for queue deletion

Change-Id: I21d9204fae006b3e0c554e15708ea2765437cd15