[sdk] Add a zircon_sysroot_for_toolchain legacy target
As of today, all consumers of the launchpad library know how to consume
it when packaged as a package separately from the sysroot except for
our Rust and Clang toolchain bots. This defines a separate sysroot
generation target for those builds so we can remove launchpad from the
primary Zircon sysroot used by Fuchsia builds and exported in the SDK.
Once the toolchain bots update to consuming a more modern SDK, we can
remove this legacy target.

Change-Id: Ic1b6e1ba09f189abc8fed204a6fb60b52d7239ad