[scripts] Reverse the logic used to find the path to the cipd binary.
If the Zircon checkout is part of a jiri checkout that includes
//buildtools then the cipd binary from there should always be used,
because we know that this cipd binary is good and of a sufficiently
recent version.

Otherwise fall back to using some unknown cipd binary found in the PATH.
This version could possibly be too old to actually work here, but if
//buildtools isn't available there is not much choice but to try.

The original intention behind find_cipd() was most likely to ensure that
cipd would be found whether //buildtools was available or not, but the
ordering of the tests happens to have the unintentional side-effect of
allowing an old cipd version in the PATH to override the up-to-date
version in //buildtools (in which case the prebuilt update will fail).

Change-Id: Ic1a31270ddd8fad2a28c5475588bf73831e76766