[debugger] Work on symbols
This works on the two ends required for symbol handling, but doesn't
connect anything yet.

On the backend, the build id is now reported with module information.
The code to read the build id was copied (with minor updates) from the
zircon inspector library.

On the frontend, a library to interface with LLVM symbols is added,
along with tracking information for modules loaded by a process. This is
not hooked up to the actual debugger yet.

Change-Id: Ia64a57b084a40490e907d13114f8570cf1f7cbb9