[debugger] Add command to get modules list.
Adds a "libs" command that lists all modules in the current process
with their load address.

Module information will be used in a followup to track currently
loaded modules dynamically.

Add a generic facility for outputting tabular information. This will
be used more widely in a followup.

Remove WeakThunk which was used only in ProcessImpl, and replace it
with a fxl::WeakPtr.

Change-Id: Ib6a41bda6dd98e2d655bcef13a8dc7274e406156