[media_player] add introspection to media_player
This CL adds the 'dump' entry to the media_player/debug directory.
The dump entry will dump the current state of the player graph in
response to command lines such as:

    cat /hub/sys/media_player/debug/dump

Due to issues with the hub, there are some limitations. First, the
/hub part of the file system isn't available using the shell on the
target machine, but a remote shell (think "fx shell") will work.
Also, the hub doesn't appear to do well with binaries of which there
are multiple instances. In order for the command above to work, the
following sequence is required:

1) run the player (e.g. set_root_view media_player_skia <file>)
   Now there are two media_player instances, a singleton that
   launches isolates and an isolate. /hub/sys shows only the
   singleton, which has no introspection.
2) terminate the player (the 'q' key terminates the skia player)
   Now only the singleton media_player instance remains. /hub/sys
   shows no media_player instances at all.
3) run the player
   Now there are two media_player instances again. This time, the
   one that shows up in /hub/sys is the new isolate.

At this point, the cat command will work.

Change-Id: Ib98ee6215e318b531f5c1feed959af9cc29e16a7