[media_player] add classes for collecting player metrics
This CL adds three classes that will be used for collecting player
metrics for reporting via introspection.

ValueTracker tracks a sequence of numeric values, determining
minimum, maxium and average values for the sequence.

RateTracker is used to determine the rate at which an event occurs.
This will be used for determining scenic frame rate.

PacketTimingTracker tracks the passage of packets through a
particular part of the player, recording packet PTS values as they
relate to the current presentation time.

Change-Id: I1015078e0f84714a4fc9c3e3d7642a3ed6cd39b3