[media][tests] Add micro-SRC tests
This CL adds a set of frequency response and SINAD tests to target our
resamplers' ability to apply micro-SRC (very slight changes in rate).
With this CL, we now run a set of audio fidelity tests at a rate ratio
(47999:48000) that represents this type of usage. This ratio joins the
other rates in our testing (48000:48000, 96000:48000, 88200:48000,
44100:48000, 24000:48000).

This CL also increases the size of mix jobs used in our tests (to 20
ms), which makes much more obvious the effects of accumulated
fractional position error in our resamplers. As a result, the measured
values for SINAD, for the 294:160 and 147:160 rate ratios, are lower.

As threshold values for the new ratio were added, and as other values
were updated because of the change in mix job size, we also applied
a reformat of frequency response values (to N.NNNNNNNe-07), to ease
future updates.

Change-Id: If881627b15a4544633affb573b4dab49c03cd0f6