[fs][fdio] Explicitly observe VFS terminating status
ManagedVfs teardown involves the following:
1) On the async dispatch thread, identiify
that the vfs is terminating.
2) Signal all open connections.
3) When all connections are terminated, invoke
a ShutdownCallback.

Unfortunately if a connection (also executing
on the async dispatch thread) receives a port
packet that their channel is "readable", it may
be dequeued without identifying that the channel
is also "signalled for teardown".

Expose the "terminating" status of the VFS layer,
and explicitly check that instead of observing the
signal. This prevents a flaky failure condition
in memfs test.

Change-Id: Iedc17ae0cce759eec81bdd5e2d5a14a92613f2fa