[debugger] Improve disassembly output.
Uses the new table output for disassembly which eliminates a lot of the
weird formatting and extra spaces.

Changes the disassembler to output a vector of items rather than
formatted output. This gives the caller more flexibililty.

Moves the OutputBuffer back to the console part of the debugger which
makes more conceptual sense.

Change-Id: Ied4d6d34e2ab1167bb65e172ebfdbf08b1a218ca