[ath10k] Add sparse_array type
The sparse_array structure is used for maintaining a mapping of
small integers to pointers. ath10k uses them as handles for tx
buffers that are shared with the firmware.

This replaces a linux "idr" structure
(https://lwn.net/Articles/103209/). Note that the interface and
functionality are designed for the needs of the ath10k driver,
but they may be a starting point for an idr drop-in replacement
if we choose to go that route.

Test: Compiles.

NET-137

Change-Id: I978587f70975056e845e8c7ba59ac805619fd4f0