[fidl][examples] Remove C++ Client unused vars
There were a couple of unused variables leftover from previous
implementations in the C++ Echo2 Client. A zx::process member variable
that was never getting initialized or used. Additionally, the
EchoClientApp::Start() method was not using the msg parameter that was
being passed in. This parameter is really only needed for the
echo2::EchoPtr::EchoString() method.

Change-Id: Ib1509d3ec0b5efdf96c958a0ed8a5c6ffeef0d0a