[go][fidl] Read go_library name from generated file
This change reads the name of the go_library for a fidl dependency from
a generated file, instead of making it the target name. This is
necessary for dealing with proper namespacing. Only the generator
backend knows what the package path is going to be for a particular
library, since it's based on the FIDL library name.

FIDL-159

Change-Id: I90926e30cef767f0857fa1a51cef57fe2d16af08