[build][arm64] Clean up arm64 boot targets
Boards that do not require a boot shim can now use a common
zircon-bootimage.bin, which contains just the kernel and bootdata.

For boards that do require a boot shim, we build a <board>-zircon-bootimage.bin,
which is zircon-bootimage.bin with the boot shim prepended to it.

Now board specific images are only created for the boards that need it.

Change-Id: Ied999511717796fd89fcd68425d315984b98142a