[debugger] Use the table printer for nouns
This improves the column-based output when listing process, threads, and
breakpoints.

Improves disassembly output by putting a max size on the parameters
which prevents long x86 parameters from pushing all of the comments
across the screen.'

Improves table column sizing by making oversize elements not force that
column to the max width. Rather, the column is the width of the largest
non-oversize item. This particularly helps disassembly byte output.

Change-Id: I7fea9cff56ed9f798f610fd4c3ab0038c040f112