[ledger] Fix race between GetObjectStatus and SetObjectStatus
GetObjectStatus is asynchronously reading 3 values from the databse and
returning with the first found instance. If a write is done between 2
reads, it was possible that no instance would be found, which is an
error.

This CL modifies the order of reads so that it is impossible to reach
this case. This is possible because the keys are ordered and if a value
is deleted, a bigger value will be written in the same transaction.

LE-461 #done

Change-Id: I635cddd4e159b196df3abd654e31209477a47483