[netstack] Packet filter: Connection state tracking
This CL implements initial state tracking for packet filtering.
The algorithm is roughly the same as the one used in OpenBSD pf,
which was based on Guido van Rooij's USENIX paper (2001).

The code also track NAT and RDR states and rewrite packets as
necessary based on the rules.

Tests:
  state_test.go
  rewritePacket_test.go

NET-513 #done

Change-Id: I43e3b2e9b95463520dc517a3bc69832dd14faa62