[dev][xdc] Replace configured_mutex and EP locks with a single lock.
We will need to process the event ring before queueing
new transfers. Since the event ring has both events
for the IN and OUT EPS, we would need to acquire
all three locks anyway.

This is as we may miss the halt register for the IN / OUT
EP being set if it is cleared before our poll loop gets to it,
so we should check the event ring to see if there were any TRBs
that had an error code which indicates a halt occurred.

Change-Id: I777a6c25bd45f870470e65dbf2738c9b005978b5