[media_player] make SetAudioRenderer accept AudioRenderer2
MediaPlayer.SetAudioRenderer used to accept an AudioRenderer and
MediaRenderer for testing purposes. This CL changes that method so it
accepts AudioRenderer2, which incorporates both AudioRenderer and
MediaRenderer functionality. AudioRenderer and MediaRenderer are
deprecated.

media_player_tests is currently disabled. A future CL will introduce
a fake AudioRenderer2 implementation, allowing media_player_tests to
be enabled.

Change-Id: If714b8675d0281f16c9e29d9e594b02b7b7d6bb9