[media_player] ensure FidlAudioRenderer doesn't stall
FidlAudioRenderer is the sink node that manages a player's
relationship to an AudioRenderer2 FIDL implementation. Previously, it
relied on very specific AudioRenderer2 behavior to prevent it from
stalling. This CL makes it more resilient to variations in
AudioRenderer2 timing behavior.

FidlAudioRenderer receives packets via the in-proc graph in which
it participates and manipulates an AudioRenderer2 implementation
accordingly. It has other responsibilities such as signalling that
end-of-stream has been reached and expressing demand for packets -
all of which has to happen within certain timing constraints.

FidlAudioRenderer previously used no timers, with one exception. That
meant that it depended on the AudioRenderer2 to do things at specific
times. For example, FidlAudioRenderer depended on
AudioRenderer2.SendPacket responses to create an opportunity for it
to check whether more packets were needed and to demand them from
upstream. If responses arrived too early, the check would occur when
there was no immediate demand. If responses arrived too late, the
check would occur after the AudioRenderer2 had starved. A timer *was*
used to update FidlAudioRenderer business when timeline transitions
were scheduled to occur.

This CL employs timers to ensure that FidlAudioRenderer obligations
are being met in a timeline manner regardless of AudioRenderer2's
timing behavior. For example, when the timeline is progressing but
demand is negative, because enough content has been delivered to
the AudioRenderer2 for the moment, FidlAudioRenderer sets a timer
for when the next packet will be required, at which time it checks
to see if it should demand another packet. Similar logic is used
when an end-of-stream marker is encountered, and FidlAudioRenderer
knows when the end-of-stream transition will occur.

Change-Id: I3805f06c299fdd8deddfe3be20c4fd69887b71c7