[debugger] Hook up symbol thread.
Adds a thread and proxy objects for interacting with the LLVM symbol
interface previously written.

Hooks up the "libs" command to refresh the module list tracked in the
symbol system.

Automatically parses the build ID mapping file "ids.txt" on startup.
Prints the information about this since not finding this file will
likely be a source of problems for users.

Adds a "sym-info" command to print the modules, build IDs, and local
file path names for the purposes of diagnosing symbol problems.

Adds a "sym-near" command to convert an address to a symbol.

Adds an "Error" type of syntax highlighting which is used to highlight
when a symbol file is not found in the summary.

The current symbol interactions are only the above explicit
interactions. Symbols will be hooked up to the stack and step commands
in a future pass.

Change-Id: I32fbf072f4ddeaa357fe52b399993adc261ba913