[operation] Add OperationContainer::Add().
This is in preparation to allow operations to be added explicitly to a
queue, instead of implicitly in the constructor for OperationBase.

Has tolerance for either method of associating an OperationBase with an
OperationContainer while we transition clients one by one.

The following syntax would be used:

container_.Add(new FooCall(...));

instead of:

new FooCall(&container_, ...);

Change-Id: If98c863fae70a7cf543555e6e5aa150ed9888f63