[dev][xdc] Handle event ring transfer events.
If we get an error, we set the transfer ring as halted (if not already),
and try to requeue and restart the transfers.

Also renamed xdc_update_configuration_state_locked to xdc_update_state_locked.
It also calls xdc_handle_events now, as this may change an EP's state,
and we need to do this before queueing a new transfer.

Change-Id: I4bec4e8af054bed05a2f4e261d7a7d88693987b0