[kernel][arm64] Override memory base not just size
If the device tree has a memory base and size, the old code would
use the size from the device tree even if the base is different than
the entry than this size will overwrite. Save and use the base address
as well. Overriding a non-zero base address with zero is not currently
supported.

Change-Id: Id1a301b84702a98abb0d41f2da44efedefaca53c