[audio][mixer] Interpolation precision
Use a modulo and denominator to communicate remaining resampling
precision to our Mixer objects, in addition to the fixed-point
frac_step_size. Because the 19.13 format could not perfectly
express arbitrary rate ratios, position error would accumulate
over time (exacerbated by large buffer sizes), leading to
distortion that manifested itself as frequency blurring.

This change improves our SINAD by as much as 20dB, depending on
resampling ratio and buffer size.

The performance hit from this addition, measured in time spent
in the Mix() function, ranges between 3-6% (for multi-channel)
to as high as 30+% for certain mono LinearSampler mixes. On
average, for stereo I/O, the impact is expected to be 15-25%.
These measurements were made on NUC.

MTWN-49 #done

Change-Id: I3bf698a98f5b8e0fa8492efdea666b9c726312d5